+++ src/edit/editgrouppage.cpp
@@ -1,0 +1,15 @@
+/*!
+ * \copyright
+ * Copyright (C) 2015 - 2020.
+ * Released under the GNU General Public License.
+ */
+
+#include "edit/editgrouppage.h"
+
+namespace cor {
+
+
+EditGroupPage::EditGroupPage(QWidget* parent, CommLayer* layer, GroupData* parser)
+    : EditPage(parent, layer, parser) {}
+
+} // namespace cor